% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_ML.R
\name{get_log_ML_bioassay}
\alias{get_log_ML_bioassay}
\title{get_log_ML_bioassay}
\usage{
get_log_ML_bioassay(
  logperms,
  successes,
  trials,
  n,
  num_trials,
  S,
  debug = FALSE
)
}
\arguments{
\item{logperms}{A vector length n containing the computed log permanents, where a zero permanent
is indicated by a -1.}

\item{successes}{A vector of length n contatining the number of successful trials at
each level.}

\item{trials}{A vector of length n containing the number of trials at each level.}

\item{n}{Sample size.}

\item{num_trials}{Number of different trials.}

\item{S}{Number of samples from the data model. That is, the number of iterations in the estimator.}

\item{debug}{If \code{TRUE}, debug information is printed.}
}
\value{
The estimated log marginal likelihood.
}
\description{
Computes the log marginal likelihood of bioassay data from the log permanents.
Given the computed log permanents logperms, this function computes the log marginal likelihood
using the formula (2.3) in [1]. It takes care of repeated trials by adding the appropriate 
log binomial factor.
}
\references{
[1] Christensen, D (2023). Inference for Bayesian nonparametric models with binary response data via permutation counting. Bayesian Analysis, Advance online publication, DOI: 10.1214/22-BA1353.
}
