% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{log_sum_exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
The log-sum-exp of the entries of the input vector.
}
\description{
Computes the log sum exp of a vector. 
Given input array = [x_1, ..., x_n], returns x_* + log(exp(x_1 - x_*) + ... + exp(x_n - x_*)),
where x_* = max(x_1, ... x_n). Ignores entries
with value -1, as these correspond to vanishing permanents.
}
\examples{
library(perms)
x = c(1,2,3,-1,-1,1)
log_sum_exp(x)
}
